Remove AttrsProxy.
